LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/update-bytecheck.https-expected.txt


FAIL Test(cors: false, main: default, imported: default) assert_array_equals: property 2, expected null but got object "[object ServiceWorker]"
PASS Test(cors: false, main: default, imported: time) 
PASS Test(cors: false, main: time, imported: default) 
PASS Test(cors: false, main: time, imported: time) 
FAIL Test(cors: true, main: default, imported: default) assert_unreached: unregister and register should not fail: NetworkError:  A network error occurred. Reached unreachable code
FAIL Test(cors: true, main: default, imported: time) assert_unreached: unregister and register should not fail: NetworkError:  A network error occurred. Reached unreachable code
FAIL Test(cors: true, main: time, imported: default) assert_unreached: unregister and register should not fail: NetworkError:  A network error occurred. Reached unreachable code
FAIL Test(cors: true, main: time, imported: time) assert_unreached: unregister and register should not fail: NetworkError:  A network error occurred. Reached unreachable code
